+++ R/embed_tooltip.R
@@ -20,7 +20,7 @@
 #' @seealso \code{\link{bs_embed_popover}}, \url{http://getbootstrap.com/javascript/#tooltips}
 #' @examples
 #' library("htmltools")
-#' tags$button(type = "button", class = "btn btn-default", "I'm a button") %>%
+#' bs_button("I'm a button") %>%
 #'   bs_embed_tooltip(title = "I'm a tooltip")
 #' @export
 #'